+++ verification.sh
@@ -1,52 +1,28 @@
 #!/bin/bash
 
-rm -rf venv_py3
-rm -rf venv_py2
+rm -rf venv_py
 
-python3 -m venv venv_py3
-source venv_py3/bin/activate
+python -m venv venv_py
+source venv_py/bin/activate
 if [ "$1" = "pypi" ]; then
-  python3 -m pip install ruuvitag_sensor
+  python -m pip install ruuvitag_sensor
 else
-  python3 -m pip install -e .
+  python -m pip install -e .
 fi
 
 if [ "$RUUVI_BLE_ADAPTER" = "Bleson" ]; then
-  python3 -m pip install git+https://github.com/TheCellule/python-bleson
+  python -m pip install git+https://github.com/TheCellule/python-bleson
 fi
 
-python3 verification.py
-ret=$?
-if [ $ret -ne 0 ]; then
-  echo 'Python 3 verification failed'
-  rm -rf venv_py3
-  exit 1
-fi
-deactivate
-rm -rf venv_py3
-
-if [ "$RUUVI_BLE_ADAPTER" = "Bleson" ]; then
-  echo 'Python 2 verification not executed for Bleson'
-  echo 'VERIFICATION COMPLETED'
-  exit 0
-fi
-
-virtualenv --python=/usr/bin/python2.7 venv_py2
-source venv_py2/bin/activate
-if [ "$1" = "pypi" ]; then
-  pip install ruuvitag_sensor
-else
-  pip install -e .
-fi
 python verification.py
 ret=$?
 if [ $ret -ne 0 ]; then
-  echo 'Python 2.7 verification failed'
-  rm -rf venv_py2
+  echo 'Python 3 verification failed'
+  rm -rf venv_py
   exit 1
 fi
 deactivate
-rm -rf venv_py2
+rm -rf venv_py
 
 echo '----------------------'
 echo 'VERIFICATION COMPLETED'