+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ### [Unreleased]
+* CHANGE: Support only Python 3.7 and greater
+* CHANGE: Updated RxPy to v4
 
 ## [1.2.1] - 2022-05-23
 * FIX: Handle too shot BLE data from RuuviTag as valida data